Moved Helpers to seperate files
move helpers to different files as they were for different purposes 

tests to check the HttpClientExtensions was creating a httpclient with headers set
